+++ src/check/subtype.rs
@@ -3,6 +3,7 @@ use crate::check::read_back::ReadBack;
 use crate::check::tcm::{TCE, TCM, TCS};
 use std::borrow::Cow;
 
+/// Check if `subtype` is the subtype of `supertype`.
 pub fn check_subtype(
     index: u32,
     (gamma, context): TCS,
@@ -10,7 +11,7 @@ pub fn check_subtype(
     supertype: Value,
 ) -> TCM<TCS> {
     let (subtype, supertype) = match (subtype, supertype) {
-        (Value::Sum((sub_tree, sub_context)), Value::Sum((mut super_tree, super_context))) => {
+        (Value::InferredSum(sub_tree), Value::Sum((mut super_tree, super_context))) => {
             for (constructor, sub_parameter) in sub_tree.into_iter() {
                 if let Some(super_parameter) = super_tree.remove(constructor.as_str()) {
                     // They're supposed to be well-typed, but I'm not sure.
@@ -18,7 +19,7 @@ pub fn check_subtype(
                     check_subtype(
                         index,
                         (Cow::Borrowed(&gamma), context.clone()),
-                        sub_parameter.eval(*sub_context.clone()),
+                        *sub_parameter,
                         super_parameter.eval(*super_context.clone()),
                     )?;
                 } else {